pyproject.toml
[project]
name = "pyldon"
version = "0.1.0"
description = "Personal Claude assistant with container-isolated agents via Matrix"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "matrix-nio[e2e]>=0.25.0",
    "aiosqlite>=0.20.0",
    "croniter>=6.0.0",
    "loguru>=0.7.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.scripts]
pyldon = "pyldon.main:main_entry"
pyldon-pair = "pyldon.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pyldon"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
